- Break goals chart into smaller components - Add min/max y/x prop support - Add `valueFormatter` prop
